hr/urls.py
"""
Router configuration for Employee and Department APIs.
"""

from rest_framework.routers import DefaultRouter

from .views import DepartmentViewSet, EmployeeViewSet

router = DefaultRouter()
router.register(r"employees", EmployeeViewSet, basename="employee")
router.register(r"departments", DepartmentViewSet)

urlpatterns = router.urls
