pyproject.toml
[project]
name = "pristine-celery-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Karthik",email = "krtk30@live.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "redis (>=6.1.0,<7.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pytest (>=8.3.5,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.11.1"
pytest-cov = "^6.1.1"
pre-commit = "^4.2.0"

[tool.pytest.ini_options]
addopts = "--cov=hr --cov-report=term-missing --cov-fail-under=95"